apps/examples/rag/requirements.txt
sentence-transformers
faiss-cpu
markdown
beautifulsoup4
requests
python-dotenv
